requirements.txt
Flask
gunicorn
Flask-Babel
psycopg2-binary
Flask-Login
Flask-WTF
Werkzeug

blinker==1.9.0
click==8.2.1
Flask==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
Werkzeug==3.1.3
WTForms==3.2.1
